added heatmap to ios

+++ GoogleMapsUtils/MPowerKit.GoogleMaps.Utils/HeatMap/WeightedLatLng.cs
@@ -1,3 +1,3 @@
 ﻿namespace MPowerKit.GoogleMaps.Utils;
 
-public record struct WeightedLatLng(Point LatLng, double Intensity = 1.0);
+public record struct WeightedLatLng(Point LatLng, float Intensity = 1f);
